[#190] refactor: 패키지명 변경
- service에서 application으
