+++ src/main/java/com/beat/global/auth/client/application/SocialService.java
@@ -1,4 +1,4 @@
-package com.beat.global.auth.client.service;
+package com.beat.global.auth.client.application;
 
 import com.beat.global.auth.client.dto.MemberInfoResponse;
 import com.beat.global.auth.client.dto.MemberLoginRequest;
